+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.4.1 - Released 20 Jan 2022
+
+### Upgrading
+
+- Add the `Parley.onRequestPermissionsResult(requestCode, permissions, grantResults);` method in the *Activity* that contains the *ParleyView* for a smooth interaction when Parley needs to request permissions. Check out the updated *Step 4* in (README.md)[README.md#step-4-forward-activity-results].
+
+### Changes:
+
+- Request camera permission if needed 
+
 ## 3.4.0 - Released 15 Nov 2021
 
 - Added option to configure where the notifications should be shown in the chat: `top` (default) or `bottom`.